+++ TaskMail/Services/ITaskHeaderService.cs
@@ -5,7 +5,7 @@ namespace TaskMailService.Services
     public interface ITaskHeaderService
     {
         List<TaskHeaderDM> GetTaskHeader(string userName,string fromDate,string toDate,out int status, out string message);
-        TaskHeaderDM InsertTaskHeader(TaskHeader taskHeaderVM, out int status, out string message, out int headerId);
-        TaskHeaderDM UpdateTaskHeader(TaskHeader taskHeaderVM, out int status, out string message, out int outHeaderId);
+        TaskHeaderDM InsertTaskHeader(TaskHeader taskHeaderVM, out int status, out string message, out long headerId);
+        TaskHeaderDM UpdateTaskHeader(TaskHeader taskHeaderVM, out int status, out string message, out long outHeaderId);
     }
 }
